requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
dependency-injector==4.0.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==2.0.0
iniconfig==1.0.1
multidict==4.7.6
packaging==20.4
pika==1.1.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.1.1
PyYAML==5.3.1
six==1.15.0
sockjs-tornado==1.0.7
toml==0.10.1
tornado==6.0.4
typing-extensions==3.7.4.3
urllib3==1.25.10
yarl==1.6.0
zipp==3.3.0
